requirements.txt
asyncpg==0.30.0
fastapi==0.115.6
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
pytest==8.3.4
SQLAlchemy==2.0.36
starlette==0.41.3
tronpy==0.5.0
uvicorn==0.34.0
